perf(lobby): isolate shared profile identity types
